refactor(auth): extract shared flow-binding, fix two follow-ups
- Move flowStatus, flowTombstone, and gcSweepEvery out of mastodon.go
  into a new flow_binding.go in the same package. Bluesky was already
  consuming all three; the implicit cross-provider coupling made the
  Bluesky implementation harder to navigate and easier to break during
  refactors. They're now explicitly shared.
- Tighten the AllowInsecurePDSURLs godoc. The flag actually disables
  scheme/internal-IP/host-consistency checks across the DID document's
  serviceEndpoint, the AS metadata's authorize/token/PAR endpoints, and
  refresh-token decoded URLs — not just the original serviceEndpoint
  scope the comment described. The userinfo guard remains
  unconditional and that's now stated explicitly.
- Replace the unreachable `return nil, fmt.Errorf("token endpoint
  exhausted DPoP-Nonce retry")` at the end of tokenRequestWithDPoP with
  `panic("unreachable: ...")`. Every loop iteration either `continue`s
  (only on attempt 0) or returns, so the previous trailing return was
  dead code with a misleading message; the panic flags any future drift
  in the loop body loudly instead of silently mis-attributing failures.
